feat: Rename project to Farmy's Ledger
Update project name and application title to reflect the new branding.
Also, refactor the app component to improve title rendering and prepare for notifications.
Update mock equipment data to be more relevant to hydroponics.
Add `Notification` type and management to `FarmContext`.
Add `HydroponicsIcon` and update `EquipmentIcon` to be more general.